requirements.txt
pytest>=2.5.0,<2.6.0
pytest-django>=2.6.0,<2.7.0
Django>=1.6.0,<1.7.0
djangorestframework>=2.3.0,<2.4.0
